requirements.txt
affine==2.4.0
attrs==25.3.0
black==25.1.0
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
click-plugins==1.1.1
cligj==0.7.2
contextily==1.6.2
contourpy==1.3.2
cycler==0.12.1
fonttools==4.58.4
geographiclib==2.0
geopandas==1.1.0
geopy==2.4.1
idna==3.10
joblib==1.5.1
kiwisolver==1.4.8
matplotlib==3.10.3
mercantile==1.2.1
mypy_extensions==1.1.0
networkx==3.5
numpy==2.3.1
osmnx==2.0.4
packaging==25.0
pandas==2.3.0
pathspec==0.12.1
pillow==11.2.1
platformdirs==4.3.8
pyogrio==0.11.0
pyparsing==3.2.3
pyproj==3.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
rasterio==1.4.3
requests==2.32.4
scikit-learn==1.7.0
scipy==1.16.0
shapely==2.1.1
six==1.17.0
threadpoolctl==3.6.0
tzdata==2025.2
urllib3==2.5.0
xyzservices==2025.4.0
